Added the implementation for DaemonManager and ManagedDaemon for Windows.
