+++ agent/engine/daemonmanager/daemon_manager_windows_test.go
@@ -1,5 +1,5 @@
-//go:build windows
-// +build windows
+//go:build windows && unit
+// +build windows,unit
 
 // Copyright Amazon.com Inc. or its affiliates. All Rights Reserved.
 //
@@ -7,18 +7,22 @@
 // not use this file except in compliance with the License. A copy of the
 // License is located at
 //
-//	http://aws.amazon.com/apache2.0/
+//      http://aws.amazon.com/apache2.0/
 //
 // or in the "license" file accompanying this file. This file is distributed
 // on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
 // express or implied. See the License for the specific language governing
 // permissions and limitations under the License.
-package csiclient
 
-import md "github.com/aws/amazon-ecs-agent/ecs-agent/manageddaemon"
+package daemonmanager
 
 const (
-	DefaultImageName      = md.EbsCsiDriver
-	DefaultSocketName     = "csi-driver.sock"
-	DefaultSocketHostPath = "C:\\ProgramData\\Amazon\\ECS\\"
+	TestDaemonName       = "TestWindowsDaemon"
+	TestImageRef         = "TestWindowsImage"
+	TestImageTag         = "testWindowsImageTag"
+	TestOtherVolumeID    = "otherTestMount"
+	TestHealthString     = "testHealth"
+	TestImagePath        = "/test/image/path"
+	TestAgentPath        = "/test/agent/path"
+	TestMountPointVolume = "testVolume"
 )
