+++ ecs-agent/manageddaemon/managed_daemon_linux_test.go
@@ -1,5 +1,5 @@
-//go:build windows
-// +build windows
+//go:build linux && unit
+// +build linux,unit
 
 // Copyright Amazon.com Inc. or its affiliates. All Rights Reserved.
 //
@@ -7,18 +7,21 @@
 // not use this file except in compliance with the License. A copy of the
 // License is located at
 //
-//	http://aws.amazon.com/apache2.0/
+//      http://aws.amazon.com/apache2.0/
 //
 // or in the "license" file accompanying this file. This file is distributed
 // on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
 // express or implied. See the License for the specific language governing
 // permissions and limitations under the License.
-package csiclient
 
-import md "github.com/aws/amazon-ecs-agent/ecs-agent/manageddaemon"
+package manageddaemon
 
 const (
-	DefaultImageName      = md.EbsCsiDriver
-	DefaultSocketName     = "csi-driver.sock"
-	DefaultSocketHostPath = "C:\\ProgramData\\Amazon\\ECS\\"
+	TestImageName                         = "TestDaemon"
+	TestImageTag                          = "testTag"
+	TestImagePath                         = "/test/image/path"
+	TestAgentPath                         = "/test/agent/path"
+	TestMountPointVolume                  = "testVolume"
+	ExpectedAgentCommunicationMountFormat = "/var/run/ecs/%s/"
+	ExpectedApplicationLogMountFormat     = "/var/log/ecs/daemons/%s/"
 )
